# Conflicts:
#	src/main/java/com/ormi5/movieblog/comment/Comment.java
#	src/main/java/com/ormi5/movieblog/comment/CommentDto.java
#	src/main/java/com/ormi5/movieblog/commentcontroller/CommentController.java
#	src/main/java/com/ormi5/movieblog/commentcontroller/CommentRepository.java
#	src/main/java/com/ormi5/movieblog/commentcontroller/CommentService.java

+++ src/main/java/com/ormi5/movieblog/postcontroller/PostController.java
@@ -34,12 +34,12 @@ public class PostController {
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<Optional<PostDto>> getPostById(@PathVariable("id") int id) {
+	public ResponseEntity<Optional<PostDto>> getPostById(@PathVariable("id") Long id) {
 		return ResponseEntity.ok(postService.getPostById(id));
 	}
 
 	@PutMapping("/{postId}")
-	public ResponseEntity<?> updatePost(@PathVariable int postId, @RequestBody PostDto postDto) {
+	public ResponseEntity<?> updatePost(@PathVariable Long postId, @RequestBody PostDto postDto) {
 		return ResponseEntity.ok(postService.updatePost(postId, postDto));
 	}
 
@@ -52,7 +52,7 @@ public class PostController {
 	 * @return result ResponseEntity
 	 */
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Optional<PostDto>> deletePostById(@PathVariable("id") int id) {
+	public ResponseEntity<Optional<PostDto>> deletePostById(@PathVariable("id") Long id) {
 		boolean deleted = postService.deletePost(id);
 		return deleted ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
 	}
